requirements.txt
apimatic-core==0.2.14
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
deprecation==2.1.0
Django==5.0.6
django-autoslug==1.9.9
django-bootstrap5==24.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.7
jsonpickle==3.0.4
jsonpointer==2.4
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
square==0.0.3
squareup==37.1.0.20240604
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
